[24.11.2021] - Рефакторинг по замечаниям - "Этого каталога не должно быть в гите.    https://github.com/filorz/orders-list-jdbc/tree/feature/add-ordering-crud/buildSrc/.gradle" - "зачем на каждый getConnection создавать новый DriverManagerDataSource ?" - "надо использовать connection pool." - "под id лучше использовать long" - "не удалось решить проблему с commit?" - "лучше так    var id = rs.getInt(1);    ordering.setId(id);    return id;" - "нужен ли тут commit?" - "почему тут ставится текущее время, а не берется значение из базы данных?" - "почему не    ordering.setId(rs.getInt("id")); ?" - "Это OrderingItemDao    тут должен быть только OrderingItem.    Ordering быть не должно" - "нет усправления транзакциями для атомарного создания заказа с позициями заказа" - "для чего нужен этот сервис?"

+++ ordersJdbc/src/main/java/ru/core/models/OrderingItem.java
@@ -2,8 +2,8 @@ package ru.core.models;
 
 public class OrderingItem {
 
-    private int id;
-    private int orderingId;
+    private long id;
+    private long orderingId;
     private String itemName;
     private int itemCount;
     private String itemPrice;
@@ -19,19 +19,19 @@ public class OrderingItem {
     public OrderingItem() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getOrderingId() {
+    public long getOrderingId() {
         return orderingId;
     }
 
-    public void setOrderingId(int orderingId) {
+    public void setOrderingId(long orderingId) {
         this.orderingId = orderingId;
     }
